e_group/requirements.txt
python-dotenv==1.0.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
supabase==2.4.2
requests==2.31.0
openai==1.27.0
python-telegram-bot==20.3
urllib3==2.1.0
